Admin Panel, Shop, Cart

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -14,21 +14,14 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('email')->unique()->nullable();
-
-             $table->string('telegram_chat_id')->unique()->nullable();
-
-             $table->string('fio_from_telegram')->default('')->nullable();
+            $table->string('telegram_chat_id')->unique()->nullable();
+            $table->string('fio_from_telegram')->default('')->nullable();
             $table->string('phone')->unique()->nullable();
-
-            // $table->string('birthday')->nullable();
-
-             $table->boolean('is_admin')->default(false);
-
-            // $table->timestamp('email_verified_at')->nullable();
-             $table->string('password')->nullable();
+            $table->boolean('is_admin')->default(false);
+            $table->string('password')->nullable();
             $table->rememberToken();
             $table->timestamps();
             $table->softDeletes();
